Remove analytics alert page
