+++ src/pages/Onboarding/OnboardingNavigator.tsx
@@ -2,7 +2,6 @@ import { createStackNavigator } from '@react-navigation/stack';
 import React from 'react';
 import { vars } from '../../utils/theme';
 import AlphaAlertOnboarding from './AlphaAlertOnboarding';
-import AnalyticsAlertOnboarding from './AnalyticsAlertOnboarding';
 import BluetoothOnboarding from './BluetoothOnboarding';
 import GetStartedOnboarding from './GetStartedOnboarding';
 import ProfileOnboarding from './ProfileOnboarding';
@@ -12,7 +11,6 @@ export type OnboardingStackParamList = {
   ProfileOnboarding: undefined;
   Bluetooth: undefined;
   AlphaAlertOnboarding: undefined;
-  AnalyticsAlertOnboarding: undefined;
 };
 
 export type OnboardingStackParamListKeys = keyof OnboardingStackParamList;
@@ -23,7 +21,6 @@ export function isOnboardingRoute(routeName: string): routeName is keyof Onboard
     'ProfileOnboarding',
     'Bluetooth',
     'AlphaAlertOnboarding',
-    'AnalyticsAlertOnboarding',
   ];
 
   return onboardingRoutes.includes(routeName as OnboardingStackParamListKeys);
@@ -46,7 +43,6 @@ const OnboardingNavigator = () => {
       <Stack.Screen name="ProfileOnboarding" component={ProfileOnboarding} />
       <Stack.Screen name="Bluetooth" component={BluetoothOnboarding} />
       <Stack.Screen name="AlphaAlertOnboarding" component={AlphaAlertOnboarding} />
-      <Stack.Screen name="AnalyticsAlertOnboarding" component={AnalyticsAlertOnboarding} />
     </Stack.Navigator>
   );
 };
